added BrowserRouter so I can use useParams, trying to add QueryClientProvider wrapper, but getting some errors, noted out for now
