+++ client/components/Homepage.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, Component } from 'react';
 import { Link, useNavigate} from 'react-router-dom'
-import FoodLog from './FoodLog';
+import NewEntry from './NewEntry';
 import SugarGraph from './SugarGraph';
 import BloodPressureGraph from './BloodPressureGraph'
 import Navbar from './Navbar'
@@ -25,11 +25,11 @@ function Homepage() {
   const usernameCookie = getCookie('username')
   console.log(usernameCookie)
   const cards = [];
-  // for (let i = 0; i < 3; i++) {
-  //   cards.push(<InfoCard key={crypto.randomUUID()}/>);
-  // }
+  
   const [buttonPopup, setButtonPopup] = useState(false);
   const [ data, setData ] = useState([]);
+  const [ lastModified, setLastModified ] = useState(Date.now())
+  const [ newData, setNewData ] = useState(Date.now())
 
   useEffect(() => {
     fetch('http://localhost:3000/api/homepage/bloodsugar')
@@ -37,34 +37,21 @@ function Homepage() {
     .then(data => {
       const array = []
       data.forEach(el => { 
-      
         if(el.username === usernameCookie){
           const dateObject = new Date(el.date);
           const options = { weekday: 'short', month: 'numeric', day: 'numeric', hour: 'numeric', minute:'numeric' };
           const formattedDate = dateObject.toLocaleString('en-US', options);
           el.date = formattedDate
-          array.push(el)
-          // const formattedData = data.map(item => {
-            // const dateObject = new Date(item.date);
-            // const options = { weekday: 'short', month: 'numeric', day: 'numeric' };
-            // const formattedDate = dateObject.toLocaleString('en-US', options);
-            // return {...item, date: formattedDate}
-          // })
-          
+          array.push(el)        
         }
-      
       })
       console.log(array)
       setData(array)
     })
     .catch(error => console.log('Error displaying entries on homepage'))
-  },[data])
+    // don't want to put a dependency, since i need it to keep displaying the updated responses, and if i put an [] it'll just fetch once
+  },[])
   
-  // const navigate = useNavigate();
-
-  // const handleClick = () => {
-  //   navigate('/foodlog');
-  // };
 
   function getCookie(cookieName) {
     const cookies = document.cookie.split('; ');
@@ -130,8 +117,17 @@ function Homepage() {
             diaPressure: formData.diaPressure,
           })
         })
+        .then(() => {
+          
+        })
+        .catch(error => console.error('Error updating data:', error))
       }
     }
+    /*state updates may be async in React, newData might not update by the time 
+    setNewData(newData+1) is called, to ensure the state updates properly must 
+    use functional form of 'setState' to guarantee we're working with the 
+    most up-to-date state and put it in .then chaining*/
+    setNewData(prevData => prevData + 1);
   }
 
   const handleChange = (e) => {
@@ -154,7 +150,7 @@ function Homepage() {
         <button id='newEntry-btn' onClick={() => setButtonPopup(true)}>New Entry</button>
       </div>
       {/* <div className='card-container'>{cards}</div> */}
-      <FoodLog trigger={buttonPopup} setTrigger={setButtonPopup} getCookie={getCookie}></FoodLog>
+      <NewEntry trigger={buttonPopup} setTrigger={setButtonPopup} getCookie={getCookie}></NewEntry>
       <div className='entriesContainer'>
         {data.map(item => (
           <div key={item._id} className='entriesHomepage'>
@@ -187,22 +183,5 @@ function Homepage() {
     </div>
   );
 }
-
-// function Homepage() {
-//   return(
-//   <div>
-//     <h1 className="hp-header">Homepage</h1>
-//     <div className="entries-container">Entries
-//       <div className="entries">
-//         <p>Date:</p>
-//         <p>Blood sugar:</p>
-//         <p>Blood Pressure:</p>
-//         <p>Time:</p>
-//         <p>Before or After meal?</p>
-//       </div>
-//     </div>
-//   </div>
-//   )
-// }
 
 export default Homepage;
