+++ client/index.js
@@ -6,7 +6,5 @@ import { BrowserRouter as Router } from 'react-router-dom';
 
 const root = ReactDom.createRoot(document.getElementById('root'));
 root.render(
-  <Router>
     <App />
-  </Router>
 );
