+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@tanstack/react-query": "^5.27.5",
     "chart.js": "^4.4.1",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
